Restore KNNInf to double.

+++ src/jCompute/Datastruct/knn/benchmark/TreeBenchmarks.java
@@ -239,11 +239,11 @@ public class TreeBenchmarks
 
 	private static void addTree(List<TreeBenchObject> objectList,KNNInf<TreeBenchObject> tree,String treeName)
 	{	
-		float[] pos;
+		double[] pos;
 		
 		for (TreeBenchObject currentObject : objectList) 
 		{
-			pos = new float[2];
+			pos = new double[2];
 			pos[0] = currentObject.getX();
 			pos[1] =  currentObject.getY();
 			tree.add(pos, currentObject);
@@ -331,7 +331,7 @@ public class TreeBenchmarks
 		
 		private void performComputation()
 		{
-			float[] pos = new float[2];
+			double[] pos = new double[2];
 			TreeBenchObject nearestObject;
 			for (TreeBenchObject currentObject : objectList) 
 			{
